+++ packages/terra-form-select/src/index.js
@@ -1,0 +1,32 @@
+import Select from './Select';
+
+import Combobox from './Combobox';
+import ComboboxField from './ComboboxField';
+import MultiSelect from './MultiSelect';
+import MultiSelectField from './MultiSelectField';
+import NativeSelect from './native-select/NativeSelect';
+import NativeSelectField from './native-select/NativeSelectField';
+import SearchSelect from './SearchSelect';
+import SearchSelectField from './SearchSelectField';
+import SelectField from './SelectField';
+import SingleSelect from './SingleSelect';
+import SingleSelectField from './SingleSelectField';
+import TagSelect from './TagSelect';
+import TagSelectField from './TagSelectField';
+
+export default Select;
+export {
+  Combobox,
+  ComboboxField,
+  MultiSelect,
+  MultiSelectField,
+  NativeSelect,
+  NativeSelectField,
+  SearchSelect,
+  SearchSelectField,
+  SelectField,
+  SingleSelect,
+  SingleSelectField,
+  TagSelect,
+  TagSelectField,
+};