+++ packages/terra-core-docs/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* Updated
+  * Updated terra list examples.
+* Changed
+  * Updated documentation for `terra-form-select`.
+
 ## 1.65.0 - (March 8, 2024)
 
 * Added